austin-rideshare-bot/requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.15.0
folium>=0.14.0
openpyxl>=3.1.0
google-generativeai>=0.6.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
langchain-core>=0.2.0
langchain-community>=0.2.0
requests>=2.31.0
python-dateutil>=2.8.2
